Service changes

+++ src/main/java/hexlet/code/service/LabelService.java
@@ -6,8 +6,11 @@ import hexlet.code.dto.label.LabelUpdateDTO;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.LabelMapper;
 import hexlet.code.repository.LabelRepository;
+import hexlet.code.util.UserUtils;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -17,6 +20,7 @@ public class LabelService {
 
     private final LabelRepository labelRepository;
     private final LabelMapper labelMapper;
+    private final UserUtils utils;
 
     public List<LabelDTO> getAll() {
         var labels = labelRepository.findAll();
@@ -34,21 +38,34 @@ public class LabelService {
     public LabelDTO create(LabelCreateDTO dto) {
         var label = labelMapper.map(dto);
         labelRepository.save(label);
-        var out = labelMapper.map(label);
-        return out;
+        return labelMapper.map(label);
     }
 
     public LabelDTO update(LabelUpdateDTO dto, Long id) {
         var label = labelRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Label with id " + id + " not found"));
+
+        checkUserPermissions();
+
         labelMapper.update(dto, label);
         labelRepository.save(label);
         return labelMapper.map(label);
     }
 
     public void destroy(Long id) {
-        labelRepository.findById(id)
+        var label = labelRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Label with id " + id + " not found"));
+
+        checkUserPermissions();
+
         labelRepository.deleteById(id);
+    }
+
+    private void checkUserPermissions() {
+        var authenticationUser = utils.getCurrentUser();
+
+        if (authenticationUser == null || !authenticationUser.getEmail().equals("hexlet@example.com")) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You do not have permission to perform this action");
+        }
     }
 }
