+++ src/main/java/hexlet/code/service/UserService.java
@@ -6,20 +6,20 @@ import hexlet.code.dto.user.UserUpdateDTO;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.UserMapper;
 import hexlet.code.repository.UserRepository;
+import hexlet.code.util.UserUtils;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class UserService {
-
     private final UserRepository userRepository;
     private final UserMapper userMapper;
-    private final PasswordEncoder passwordEncoder;
-
+    private final UserUtils utils;
 
     public UserDTO show(Long id) {
         var user = userRepository.findById(id)
@@ -35,20 +35,31 @@ public class UserService {
     public UserDTO create(UserCreateDTO dto) {
         var user = userMapper.map(dto);
         userRepository.save(user);
-
         return userMapper.map(user);
     }
 
     public UserDTO update(UserUpdateDTO dto, Long id) {
         var user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
-        userMapper.update(dto, user);
 
+        var authenticationUser = utils.getCurrentUser();
+        if (authenticationUser == null || !authenticationUser.getId().equals(user.getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You do not have permission to update this account");
+        }
+
+        userMapper.update(dto, user);
         userRepository.save(user);
         return userMapper.map(user);
     }
 
     public void destroy(Long id) {
+        var user = userRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
+
+        var authenticationUser = utils.getCurrentUser();
+        if (authenticationUser == null || !authenticationUser.getId().equals(user.getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You do not have permission to delete this account");
+        }
         userRepository.deleteById(id);
     }
 }
